Filter doctypes from Jade
In-file doctype tags are unnecessary in the build system as set up. At
best, they are redundant because we already specify doctype as a
parameter to the compiler, and Jade is smart enough to insert a doctype
in the output where needed if we omit it. At worst, the in-file tags
break AngularJS by emitting doctype tags in partials.

However, for better tooling support, our Jade files need to document
what doctype they are. As a solution, this change causes the Gulpfile to
strip the doctype lines from the Jade files before passing them to the
compiler, allowing us to include doctypes in the files for tooling
support while ignoring them when compiling.